Implement NeoForge support
