+++ common/src/main/java/eu/midnightdust/midnightcontrols/client/virtualkeyboard/clickhandler/AbstractScreenClickHandler.java
@@ -1,0 +1,7 @@
+package eu.midnightdust.midnightcontrols.client.virtualkeyboard.clickhandler;
+
+import net.minecraft.client.gui.screen.Screen;
+
+public abstract class AbstractScreenClickHandler<T extends Screen> {
+    public abstract void handle(T screen, double mouseX, double mouseY);
+}
