+++ neoforge/src/main/java/eu/midnightdust/midnightcontrols/neoforge/MidnightControlsClientNeoforge.java
@@ -5,6 +5,7 @@ import eu.midnightdust.midnightcontrols.client.MidnightControlsConfig;
 import eu.midnightdust.midnightcontrols.client.util.platform.NetworkUtil;
 import eu.midnightdust.midnightcontrols.packet.ControlsModePayload;
 import eu.midnightdust.midnightcontrols.packet.HelloPayload;
+import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.resource.DirectoryResourcePack;
 import net.minecraft.resource.ResourcePackInfo;
 import net.minecraft.resource.ResourcePackPosition;
@@ -21,6 +22,7 @@ import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.client.event.ClientPlayerNetworkEvent;
 import net.neoforged.neoforge.client.event.ClientTickEvent;
 import net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent;
+import net.neoforged.neoforge.client.event.ScreenEvent;
 import net.neoforged.neoforge.event.AddPackFindersEvent;
 import net.neoforged.neoforgespi.locating.IModFile;
 
@@ -34,6 +36,8 @@ import static eu.midnightdust.midnightcontrols.client.MidnightControlsClient.BIN
 import static eu.midnightdust.midnightcontrols.client.MidnightControlsClient.BINDING_LOOK_UP;
 import static eu.midnightdust.midnightcontrols.client.MidnightControlsClient.BINDING_RING;
 import static eu.midnightdust.midnightcontrols.client.MidnightControlsClient.client;
+import static eu.midnightdust.midnightcontrols.client.MidnightControlsClient.clickInterceptor;
+
 
 @Mod(value = NAMESPACE, dist = Dist.CLIENT)
 public class MidnightControlsClientNeoforge {
@@ -89,6 +93,16 @@ public class MidnightControlsClientNeoforge {
         @SubscribeEvent
         public static void startClientTick(ClientTickEvent.Pre event) {
             MidnightControlsClient.onTick(client);
+        }
+        @SubscribeEvent
+        public static void onMouseButtonPressed(ScreenEvent.MouseButtonPressed.Pre event) {
+            if (!event.isCanceled()) {
+                Screen screen = event.getScreen();
+                double mouseX = event.getMouseX();
+                double mouseY = event.getMouseY();
+
+                clickInterceptor.intercept(screen, mouseX, mouseY);
+            }
         }
     }
 }
